+++ src/components/map/navigraph/NavigraphProcedures.vue
@@ -3,43 +3,24 @@
 </template>
 
 <script setup lang="ts">
-import { useStore } from '~/store';
 import { Feature } from 'ol';
 import type { ShallowRef } from 'vue';
 import type VectorSource from 'ol/source/Vector';
-import { useMapStore } from '~/store/map';
 import { LineString, Point } from 'ol/geom';
 import type { Coordinate } from 'ol/coordinate';
 import greatCircle from '@turf/great-circle';
 import type { NavigraphNavDataAirportWaypoint } from '~/utils/backend/navigraph/navdata/types';
+import type { DataStoreNavigraphProcedure, DataStoreNavigraphProceduresAirport } from '~/composables/data';
 
 defineSlots<{ default: () => any }>();
 
 const source = inject<ShallowRef<VectorSource>>('navigraph-source');
 
-const store = useStore();
-const mapStore = useMapStore();
 const dataStore = useDataStore();
 
 let features: Feature[] = [];
 
-function constructWaypointGeometries(coordinates: Coordinate[]): Coordinate[] {
-    const result: Coordinate[] = [];
-
-    for (let i = 0; i < coordinates.length; i++) {
-        const coordinate = coordinates[i];
-        const nextCoordinate = coordinates[i + 1];
-        if (!nextCoordinate) continue;
-
-        const circle = greatCircle(coordinate, nextCoordinate, { npoints: 2 });
-        if (circle.geometry.type === 'LineString') result.push(...circle.geometry.coordinates);
-        if (circle.geometry.type === 'MultiLineString') result.push(...circle.geometry.coordinates.flatMap(x => x));
-    }
-
-    return result;
-}
-
-function addWaypoints(newFeatures: Feature[], waypoints: NavigraphNavDataAirportWaypoint[], constraints: boolean, procedure: string) {
+function addWaypoints(newFeatures: Feature[], waypoints: NavigraphNavDataAirportWaypoint[], constraints: boolean, procedure: string, name?: string) {
     for (let i = 0; i < waypoints.length; i++) {
         const waypoint = waypoints[i];
         const nextWaypoint = waypoints[i + 1];
@@ -55,13 +36,8 @@ function addWaypoints(newFeatures: Feature[], waypoints: NavigraphNavDataAirport
             geometry: new LineString(coords),
             dataType: 'navdata',
             procedure,
+            name,
             type: 'enroute',
-
-            altitude: nextWaypoint.altitude,
-            altitude1: nextWaypoint.altitude1,
-            altitude2: nextWaypoint.altitude2,
-            speed: nextWaypoint.speed,
-            speedLimit: nextWaypoint.speedLimit,
         }));
 
         if (constraints) {
@@ -72,26 +48,51 @@ function addWaypoints(newFeatures: Feature[], waypoints: NavigraphNavDataAirport
                 waypoint: x.identifier,
                 ref: x.ref,
                 type: 'enroute-waypoint',
+
+                altitude: x.altitude,
+                altitude1: x.altitude1,
+                altitude2: x.altitude2,
+                speed: x.speed,
+                speedLimit: x.speedLimit,
             })));
         }
     }
 }
 
-watch([dataStore.navigraphProcedures.sids], () => {
+function processSidOrStar(newFeatures: Feature[], { procedure: { waypoints, transitions: { enroute, runway }, procedure: { identifier } }, constraints, transitions }: DataStoreNavigraphProcedure, type: 'sid' | 'star', runways: string[]) {
+    addWaypoints(newFeatures, waypoints, constraints, type, identifier);
+
+    const runwayTransitions = runway.filter(x => !runways.length || runways.some(y => y === x.name));
+
+    if (runwayTransitions.length) {
+        addWaypoints(newFeatures, runwayTransitions.flatMap(x => x.waypoints), constraints, type);
+    }
+
+    const enrouteTransitions = enroute.filter(x => !transitions.length || transitions.some(y => y === x.name));
+    addWaypoints(newFeatures, enrouteTransitions.flatMap(x => x.waypoints), constraints, type);
+}
+
+watch(() => dataStore.navigraphProcedures, () => {
     const newFeatures: Feature[] = [];
 
-    for (const { procedure: { waypoints, transitions: { enroute, runway } }, constraints, runway: activeRunway, transition } of Object.values({ ...dataStore.navigraphProcedures.sids.value, ...dataStore.navigraphProcedures.stars.value })) {
-        addWaypoints(newFeatures, waypoints, constraints, 'sid');
-
-        const runwayTransitions = !activeRunway ? runway : runway.filter(x => x.name === activeRunway);
-
-        if (runwayTransitions.length) {
-            addWaypoints(newFeatures, runwayTransitions.flatMap(x => x.waypoints), constraints, 'sid');
+    for (const { sids, stars, approaches, runways } of Object.values(dataStore.navigraphProcedures as Record<string, DataStoreNavigraphProceduresAirport>)) {
+        for (const item of Object.values(sids)) {
+            processSidOrStar(newFeatures, item, 'sid', runways);
         }
 
-        const enrouteTransition = transition && enroute.find(x => x.name === transition);
-        if (enrouteTransition) addWaypoints(newFeatures, enrouteTransition.waypoints, constraints, 'sid');
-        else if (enroute.length) addWaypoints(newFeatures, enroute.flatMap(x => x.waypoints), constraints, 'sid');
+        for (const item of Object.values(stars)) {
+            processSidOrStar(newFeatures, item, 'star', runways);
+        }
+
+        for (const item of Object.values(approaches)) {
+            const { procedure: { waypoints, transitions: approachTransitions, procedure: { missedApproach } }, constraints, transitions } = item;
+            addWaypoints(newFeatures, waypoints, constraints, 'approaches');
+
+            const enrouteTransitions = approachTransitions.filter(x => !transitions.length || transitions.some(y => y === x.name));
+            addWaypoints(newFeatures, enrouteTransitions.flatMap(x => x.waypoints), constraints, 'approaches');
+
+            if (missedApproach.length) addWaypoints(newFeatures, missedApproach, constraints, 'missedApproach');
+        }
     }
 
     source?.value.removeFeatures(features);
@@ -99,7 +100,7 @@ watch([dataStore.navigraphProcedures.sids], () => {
     source?.value.addFeatures(features);
 }, {
     immediate: true,
-    deep: true,
+    deep: 4,
 });
 
 onBeforeUnmount(() => {